removed leading and trailing whitespace from keys
